@@ -11,25 +11,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.8",
-    "@angular/common": "^6.1.8",
-    "@angular/compiler": "^6.1.8",
-    "@angular/core": "^6.1.8",
-    "@angular/forms": "^6.1.8",
-    "@angular/http": "^6.1.8",
-    "@angular/platform-browser": "^6.1.8",
-    "@angular/platform-browser-dynamic": "^6.1.8",
-    "@angular/router": "^6.1.8",
-    "angularfire2": "^5.0.2",
-    "bootstrap": "^4.1.2",
+    "@angular/animations": "~8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "~8.0.0",
+    "@angular/compiler": "~8.0.0",
+    "@angular/core": "~8.0.0",
+    "@angular/forms": "~8.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "~8.0.0",
+    "@angular/platform-browser-dynamic": "~8.0.0",
+    "@angular/router": "~8.0.0",
+    "angularfire2": "^5.1.2",
     "core-js": "^2.5.4",
-    "firebase": "^5.5.1",
-    "hoek": "^5.0.4",
+    "firebase": "^6.1.0",
+    "hammerjs": "^2.0.8",
+    "ngx-cookie-service": "^2.1.0",
     "parse-ms": "^2.0.0",
-    "pretty-ms": "^3.2.0",
-    "querystring": "^0.2.0",
-    "rxjs": "^6.3.2",
-    "zone.js": "^0.8.26"
+    "pretty-ms": "^4.0.0",
+    "rxjs": "~6.5.2",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.6.8",
