@@ -5,13 +5,14 @@ import {SpotifyService} from '../../services/spotify/spotify.service';
 import {PlaylistService} from '../../services/playlist/playlist.service';
 import {StatusBarService} from '../../services/status-bar/status-bar.service';
 import {animate, state, style, transition, trigger} from '@angular/animations';
-import {MatDialog, MatDialogConfig} from '@angular/material';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import {NewPlaylistDialogComponent} from '../new-playlist-dialog/new-playlist-dialog.component';
 import {Router} from '@angular/router';
 import { SpotifyPlaylistRespose } from 'src/app/interfaces/playlist/spotifyPlaylistResponse.interface';
 import { CurrentTrack } from 'src/app/interfaces/track/current-track.interface';
-import { SpotifyToken } from 'src/app/interfaces/spotify-token/spotify-token.interface';
+// import { SpotifyToken } from 'src/app/interfaces/spotify-token/spotify-token.interface';
 import { PlaylistData } from 'src/app/interfaces/playlist/spotfiy-playlist-data.interface';
+// import { PlaylistData } from 'src/app/interfaces/playlist/playlist-data.interface';
 import { SpotifyPlaybackService } from 'src/app/services/spotify-playback/spotify-playback.service';
 import { AuthService } from 'src/app/services/auth/auth.service';
 import { UtilService } from 'src/app/services/util/util.service';
