@@ -2,10 +2,8 @@ import {EventEmitter, Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders, HttpParams} from '@angular/common/http';
 import {stringify} from 'query-string';
 import {ActivatedRoute, Router} from '@angular/router';
-import {AngularFirestore} from 'angularfire2/firestore';
 import {concat, of} from 'rxjs';
 import {switchMap} from 'rxjs/internal/operators';
-import {AngularFireAuth} from 'angularfire2/auth';
 import {TrackService} from '../track/track.service';
 import {PlaylistService} from '../playlist/playlist.service';
 import {environment} from '../../../environments/environment';
@@ -20,8 +18,6 @@ export class SpotifyService {
   private didPlaylistChange: boolean;
 
   constructor(private _http: HttpClient, private router: Router,
-              private afs: AngularFirestore,
-              private afa: AngularFireAuth,
               private trackService: TrackService,
               private route: ActivatedRoute,
               private playlistService: PlaylistService,
@@ -31,13 +27,6 @@ export class SpotifyService {
     this.playlistService.didPlaylistChange$.subscribe(value => this.didPlaylistChange = value);
     this.didPlaylistChange = false;
   }
-
-  // getLoginURI() {
-  //   return this.loginURI;
-  // }
-  // login() {
-  //    window.location.href = this.getLoginURI();
-  // }
 
   getUser() {
     const httpOptions = {
@@ -49,59 +38,14 @@ export class SpotifyService {
     return this._http.get(url, httpOptions);
   }
 
-  storeToken(token) {
-    const expireTime = new Date();
-    expireTime.setMinutes(expireTime.getMinutes() + (token['expires_in'] / 60));
-    this.afa.user
-      .subscribe(user => {
-        if (user) {
-          this.afs.collection('users').doc(user['email']).update({spotifyToken: {token: token.access_token, expires: String(expireTime)}});
-        }
-    });
-    this.router.navigate(['/']);
-  }
-
-  getAuthToken() {
-    return this.afa.user.pipe(
-      switchMap(user => {
-        if (user) {
-          return this.afs.collection('users').doc(user['email']).valueChanges();
-        } else {
-          return of();
-        }
-      }),
-      switchMap(token => {
-        const currentTime = new Date();
-        if (token) {
-          // @ts-ignore
-          if (token.spotifyToken.expires < String(currentTime)) {
-            this.logout();
-          }
-          return [token['spotifyToken']];
-        } else {
-          return of();
-        }
-      })
-    );
-  }
 
   getAllPlaylists(morePlaylists?) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${this.cookieService.get('spotifyToken')}`
-      })
-    };
     const url = morePlaylists ? morePlaylists : this.spotifyApiBaseURI + '/me/playlists?limit=50';
-    return this._http.get(url, httpOptions);
+    return this._http.get(url);
   }
 
-  getTrack(token, id) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    };
-    return this._http.get(this.spotifyApiBaseURI + `/tracks/${id}`, httpOptions);
+  getTrack(id) {
+    return this._http.get(this.spotifyApiBaseURI + `/tracks/${id}`);
   }
 
   getTracksFromPlaylist(playlistID, offset, limit) {
@@ -137,22 +81,12 @@ export class SpotifyService {
     return array;
   }
 
-  addTracksToPlaylist(token, owner, playlistID, tracks) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    };
-    return this._http.put(this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/tracks`, tracks, httpOptions);
+  addTracksToPlaylist(owner, playlistID, tracks) {
+    return this._http.put(this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/tracks`, tracks);
   }
 
-  postTracksToPlaylist(token, owner, playlistID, tracks) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    };
-    return this._http.post(this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/tracks`, tracks, httpOptions);
+  postTracksToPlaylist(owner, playlistID, tracks) {
+    return this._http.post(this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/tracks`, tracks);
   }
 
   /*
@@ -161,132 +95,67 @@ export class SpotifyService {
       - figure out logic to pull the next 100 tracks once you are at last track in last batch
       - same point as above but for previous tracks
   */
-  playSpotifyTrack(token, tracks,  offset, deviceID, currentPosition) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
+  playSpotifyTrack(tracks,  offset, deviceID, currentPosition) {
     return this._http.put(this.spotifyApiBaseURI + `/me/player/play?${deviceID}`,
-    {uris: tracks, position_ms: currentPosition, offset: {position: offset}}, httpOptions);
+    {uris: tracks, position_ms: currentPosition, offset: {position: offset}});
   }
 
-  pauseSpotifyTrack(token, deviceID) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.put(this.spotifyApiBaseURI + `/me/player/pause?${deviceID}`, {}, httpOptions);
+  pauseSpotifyTrack(deviceID) {
+    return this._http.put(this.spotifyApiBaseURI + `/me/player/pause?${deviceID}`, {});
   }
 
-  setRepeatMode(token, context, deviceID) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.put(this.spotifyApiBaseURI + `/me/player/repeat?state=${context}&device_id${deviceID}`, {}, httpOptions);
+  setRepeatMode(context, deviceID) {
+    return this._http.put(this.spotifyApiBaseURI + `/me/player/repeat?state=${context}&device_id${deviceID}`, {});
   }
 
-  changeSpotifyTrackVolume(token, deviceID, volume) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.put(this.spotifyApiBaseURI + `/me/player/volume?volume_percent=${volume}&device_id=${deviceID}`, {}, httpOptions);
+  changeSpotifyTrackVolume(deviceID, volume) {
+    return this._http.put(this.spotifyApiBaseURI + `/me/player/volume?volume_percent=${volume}&device_id=${deviceID}`, {});
   }
 
-  getCurrentPlayer(token) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.get(this.spotifyApiBaseURI + `/me/player`, httpOptions);
+  getCurrentPlayer() {
+    return this._http.get(this.spotifyApiBaseURI + `/me/player`);
   }
 
-  getAvailableDevices(token) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.get(this.spotifyApiBaseURI + `/me/player/devices`, httpOptions);
+  getAvailableDevices() {
+    return this._http.get(this.spotifyApiBaseURI + `/me/player/devices`);
   }
 
-  getCurrentSong(token) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.get(this.spotifyApiBaseURI + `/me/player/currently-playing`, httpOptions);
+  getCurrentSong() {
+    return this._http.get(this.spotifyApiBaseURI + `/me/player/currently-playing`);
   }
 
-  createPlaylist(token, body) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.post(this.spotifyApiBaseURI + `/me/playlists`, body, httpOptions);
+  createPlaylist(body) {
+    return this._http.post(this.spotifyApiBaseURI + `/me/playlists`, body);
   }
 
-  uploadPlaylistCover(token: string, image: File, owner: string, playlistID: string) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'image/jpeg'
-      })
-    };
-    return this._http.put(this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/images`, image, httpOptions);
+  uploadPlaylistCover(image: File, owner: string, playlistID: string) {
+    return this._http.put(this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/images`, image);
   }
+  // TODO: fix function to not have to use token
+  // createNewPlaylist(body, image) {
+  //   let authToken = '';
+  //   return this.getAuthToken()
+  //     .pipe(
+  //       switchMap((token: string) => {
+  //         authToken = token['token'];
+  //         return this.createPlaylist(authToken, body);
+  //       }),
+  //       switchMap((data: Object) => this.uploadPlaylistCover(authToken, image, data['owner']['id'], data['id']))
+  //     );
+  // }
 
-  createNewPlaylist(body, image) {
-    let authToken = '';
-    return this.getAuthToken()
-      .pipe(
-        switchMap((token: string) => {
-          authToken = token['token'];
-          return this.createPlaylist(authToken, body);
-        }),
-        switchMap((data: Object) => this.uploadPlaylistCover(authToken, image, data['owner']['id'], data['id']))
-      );
-  }
-
-  getUsersSavedAlbums(token, moreAlbums?) {
+  getUsersSavedAlbums(moreAlbums?) {
     const url = moreAlbums ? moreAlbums : this.spotifyApiBaseURI + `/me/albums`;
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      })
-    };
-    return this._http.get(url, httpOptions);
+    return this._http.get(url);
   }
 
-  replaceTrack(token: string, owner: string, playlistID: string, startIndex: number, endIndex: number) {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    };
+  replaceTrack(owner: string, playlistID: string, startIndex: number, endIndex: number) {
     const requestBody = {
       range_start: startIndex,
       insert_before: endIndex
     };
     const url = this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/tracks`;
-    return this._http.put(url, requestBody, httpOptions);
+    return this._http.put(url, requestBody);
   }
 
   mapTrackURIs(tracks) {
@@ -294,131 +163,115 @@ export class SpotifyService {
       return track.track.uri;
     });
   }
+  // TODO: fix function to not have to use token
+  // shuffler(playlistID, tracks) {
+  //   let authToken = '';
+  //   let playlistInfo = {};
+  //   this.getAuthToken()
+  //     .pipe(
+  //       switchMap((token: string) => {
+  //         authToken = token['token'];
+  //         return this.playlistService.getSavedPlaylist(playlistID);
+  //       }),
+  //       switchMap(playlistData => {
+  //         const originalTracks = [];
+  //         const shuffledPlaylist = [];
+  //         playlistInfo = playlistData;
+  //         for (const t in tracks) {
+  //           if (t) {
+  //             originalTracks.push(tracks[t]);
+  //           }
+  //         }
+  //         const shuffledTracks = this.shuffleTracks(tracks);
+  //         for (let i = 0; i < originalTracks.length; i++) {
+  //           for (let j = 0; j < shuffledTracks.length; j++) {
+  //             if (originalTracks[i] === shuffledTracks[j]) {
+  //               const playlist = {
+  //                 name: playlistData['name'],
+  //                 owner: {
+  //                   id: playlistData['id']
+  //                 },
+  //                 id: playlistData['playlistID'],
+  //                 tracks: {
+  //                   total: playlistData['playlistLength']
+  //                 }
+  //               };
+  //               const tuple = {
+  //                 range_start: i,
+  //                 insert_before: j,
+  //                 name: originalTracks[i]['track']['name']
+  //               };
+  //               // shuffledPlaylist.push(this.replaceTrack(authToken, playlistData['owner'], playlistData['playlistID'], i, j));
+  //               this.playlistService.didPlaylistChange$.emit(true);
+  //               // this.playlistService.savePlaylist(playlist, tuple);
+  //               // return this.replaceTrack(authToken, playlistData['owner'], playlistData['playlistID'], i, j);
+  //             }
+  //           }
+  //         }
+  //         return concat(...[]);
+  //       })
+  //     ).subscribe((data) => {
+  //       this.playlistService.didPlaylistChange$.emit(false);
+  //     });
+  // }
 
-  shuffler(playlistID, tracks) {
-    let authToken = '';
-    let playlistInfo = {};
-    this.getAuthToken()
-      .pipe(
-        switchMap((token: string) => {
-          authToken = token['token'];
-          return this.playlistService.getSavedPlaylist(playlistID);
-        }),
-        switchMap(playlistData => {
-          const originalTracks = [];
-          const shuffledPlaylist = [];
-          playlistInfo = playlistData;
-          for (const t in tracks) {
-            if (t) {
-              originalTracks.push(tracks[t]);
-            }
-          }
-          const shuffledTracks = this.shuffleTracks(tracks);
-          for (let i = 0; i < originalTracks.length; i++) {
-            for (let j = 0; j < shuffledTracks.length; j++) {
-              if (originalTracks[i] === shuffledTracks[j]) {
-                const playlist = {
-                  name: playlistData['name'],
-                  owner: {
-                    id: playlistData['id']
-                  },
-                  id: playlistData['playlistID'],
-                  tracks: {
-                    total: playlistData['playlistLength']
-                  }
-                };
-                const tuple = {
-                  range_start: i,
-                  insert_before: j,
-                  name: originalTracks[i]['track']['name']
-                };
-                // shuffledPlaylist.push(this.replaceTrack(authToken, playlistData['owner'], playlistData['playlistID'], i, j));
-                this.playlistService.didPlaylistChange$.emit(true);
-                // this.playlistService.savePlaylist(playlist, tuple);
-                // return this.replaceTrack(authToken, playlistData['owner'], playlistData['playlistID'], i, j);
-              }
-            }
-          }
-          return concat(...[]);
-        })
-      ).subscribe((data) => {
-        this.playlistService.didPlaylistChange$.emit(false);
-      });
-  }
+  // TODO: fix function to not have to use token
+  // addShuffledTracksToPlaylist(playlistID, tracks) {
+  //   let authToken = '';
+  //   const that = this;
+  //   const trackURIs = this.mapTrackURIs(tracks);
+  //   let amountOfLoops = 0;
+  //   let playlistInfo = {};
+  //   return this.getAuthToken()
+  //     .pipe(
+  //       switchMap(token => {
+  //         authToken = token['token'];
+  //         return that.playlistService.getSavedPlaylist(playlistID);
+  //       }),
+  //       switchMap(data => {
+  //         playlistInfo = data;
+  //         return that.addTracksToPlaylist(authToken, playlistInfo['owner'], playlistInfo['playlistID'], {uris: []});
+  //       }),
+  //       switchMap(() => {
+  //         const tempList = [];
+  //         amountOfLoops = Math.ceil(playlistInfo['playlistLength'] / 100);
+  //         for (let i = amountOfLoops - 1; i >= 0; i--) {
+  //           const tt = trackURIs.slice(i * 100, ((i + 1) * 100));
+  //           tempList.unshift(that.postTracksToPlaylist(authToken, playlistInfo['owner'], playlistInfo['playlistID'], {uris: tt}));
+  //         }
+  //         return concat(...tempList);
+  //       })
+  //     );
+  // }
 
-  addShuffledTracksToPlaylist(playlistID, tracks) {
-    let authToken = '';
-    const that = this;
-    const trackURIs = this.mapTrackURIs(tracks);
-    let amountOfLoops = 0;
-    let playlistInfo = {};
-    return this.getAuthToken()
-      .pipe(
-        switchMap(token => {
-          authToken = token['token'];
-          return that.playlistService.getSavedPlaylist(playlistID);
-        }),
-        switchMap(data => {
-          playlistInfo = data;
-          return that.addTracksToPlaylist(authToken, playlistInfo['owner'], playlistInfo['playlistID'], {uris: []});
-        }),
-        switchMap(() => {
-          const tempList = [];
-          amountOfLoops = Math.ceil(playlistInfo['playlistLength'] / 100);
-          for (let i = amountOfLoops - 1; i >= 0; i--) {
-            const tt = trackURIs.slice(i * 100, ((i + 1) * 100));
-            tempList.unshift(that.postTracksToPlaylist(authToken, playlistInfo['owner'], playlistInfo['playlistID'], {uris: tt}));
-          }
-          return concat(...tempList);
-        })
-      );
-  }
-
-  removeTracks(token, owner, playlistID, tracks) {
+  removeTracks(owner, playlistID, tracks) {
     const options = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      }),
+      headers: new HttpHeaders({}),
       body: tracks
     };
     return this._http.delete(
       this.spotifyApiBaseURI + `/users/${owner}/playlists/${playlistID}/tracks`, options);
   }
 
-  removeDuplicateTracks(playlistID, tracks) {
-    const that = this;
-    let authToken = '';
-    let playlistInfo = {};
-    this.getAuthToken()
-      .pipe(
-        switchMap(token => {
-          authToken = token['token'];
-          return that.playlistService.getSavedPlaylist(playlistID);
-        }),
-        switchMap(data => {
-          playlistInfo = data;
-          return that.removeTracks(authToken, playlistInfo['owner'], playlistInfo['playlistID'], {tracks});
-        }),
-      ).subscribe(() => {});
-  }
+  // TODO: fix function to not have to use token
+  // removeDuplicateTracks(playlistID, tracks) {
+  //   const that = this;
+  //   let authToken = '';
+  //   let playlistInfo = {};
+  //   this.getAuthToken()
+  //     .pipe(
+  //       switchMap(token => {
+  //         authToken = token['token'];
+  //         return that.playlistService.getSavedPlaylist(playlistID);
+  //       }),
+  //       switchMap(data => {
+  //         playlistInfo = data;
+  //         return that.removeTracks(authToken, playlistInfo['owner'], playlistInfo['playlistID'], {tracks});
+  //       }),
+  //     ).subscribe(() => {});
+  // }
 
-  checkSavedTrack(token, trackIDs) {
-    const options = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    };
-
-    return this._http.get(
-      this.spotifyApiBaseURI + `/me/tracks/contains?ids=${trackIDs}`, options);
-  }
-
-  logout() {
-    this.afa.user.subscribe(user => {
-      if (user) {
-        this.afs.collection('users').doc(user['email']).update({spotifyToken: {token: '', expires: null}})
-          .then(() => this.playlistService.deletePlaylists());
-      }
-    });
+  checkSavedTrack(trackIDs) {
+    return this._http.get(this.spotifyApiBaseURI + `/me/tracks/contains?ids=${trackIDs}`);
   }
 }
